aiogram-multibot/requirements.txt
aiogram==3.6
charset-normalizer==2.1.1
idna==3.4
python-dotenv==0.21.1
pytz==2022.7.1
